debugging

+++ train.py
@@ -92,9 +92,9 @@ def train(nn_model,train_set_loader,val_set_loader,loss_func,optimizer, config):
         for batch_id,(image,gt) in enumerate(train_set_loader):
             
             image = image.squeeze(1)
-            
+            image = image.float()
             gt = gt.squeeze(1)
-            
+            gt = gt.float()
             nn_model.train()
             if(configuration.training_config.device.type == 'cuda'):
                 image,gt = image.to(device=configuration.training_config.device.type,dtype=torch.int64),gt.to(device=configuration.training_config.device.type,dtype=torch.int64)
@@ -102,7 +102,7 @@ def train(nn_model,train_set_loader,val_set_loader,loss_func,optimizer, config):
                 image,gt = image, gt
             
             
-            output = nn_model(image.float())
+            output = nn_model(image)
             loss = loss_func(output, gt)
